// apps/web/src/payload-types.ts
export interface Author {
  id: number;
  name: string;
  slug: string;
  bio?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "topics".
 */
export interface Topic {
  id: number;
  name: string;
  slug: string;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number;
  alt: string;
  sha256: string;
  assetVersion: string;
  durationSeconds: number;
  license: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "courses".
 */
export interface Course {
  id: number;
  title: string;
  slug: string;
  description: string;
  author: number | Author;
  topic: number | Topic;
  cover: 'window' | 'law' | 'trade' | 'choice';
  introduction: string;
  level: 'Beginner' | 'Intermediate';
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "lessons".
 */
export interface Lesson {
  id: number;
  title: string;
  slug: string;
  description: string;
  author: number | Author;
  topic: number | Topic;
  cover: 'window' | 'law' | 'trade' | 'choice';
  course: number | Course;
  order: number;
  format: 'audio' | 'video';
  asset: number | Media;
  transcript: string;
  sources?:
    | {
        title: string;
        url: string;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "articles".
 */
export interface Article {
  id: number;
  title: string;
  slug: string;
  description: string;
  author: number | Author;
  topic: number | Topic;
  cover: 'window' | 'law' | 'trade' | 'choice';
  body: string;
  readingMinutes: number;
  sources?:
    | {
        title: string;
        url: string;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}